requirements.txt
Django==5.0.3
pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
pytorch-lightning==2.2.1